+++ R/grpreg.R
@@ -23,7 +23,21 @@ grpreg <- function(X, y, group=1:ncol(X), penalty=c("grLasso", "grMCP", "grSCAD"
   if (gamma <= 2 & penalty=="grSCAD") stop("gamma must be greater than 2 for the SCAD penalty", call.=FALSE)
   if (nlambda < 2) stop("nlambda must be at least 2", call.=FALSE)
   if (alpha > 1 | alpha <= 0) stop("alpha must be in (0, 1]", call.=FALSE)
-
+  
+  # Check for expandedMatix
+  if(inherits(X, "expandedMatrix")) {
+    expanded <- TRUE
+    group <- X$group
+    knots <- X$knots
+    boundary <- X$boundary
+    degree <- X$degree
+    originalx <- X$originalx
+    type <- X$type
+    X <- X$X
+  } else {
+    expanded <- FALSE
+  }
+  
   # Construct XG, yy
   bilevel <- strtrim(penalty, 2) != "gr"
   yy <- newY(y, family)
@@ -55,16 +69,18 @@ grpreg <- function(X, y, group=1:ncol(X), penalty=c("grLasso", "grMCP", "grSCAD"
     if (bilevel) fit <- .Call("lcdfit_gaussian", XG$X, yy, penalty, K1, K0, lambda, alpha, eps, 0, gamma, tau, as.integer(max.iter), XG$m, as.integer(dfmax), as.integer(gmax), as.integer(user.lambda))
     else fit <- .Call("gdfit_gaussian", XG$X, yy, penalty, K1, K0, lambda, lam.max, alpha, eps, as.integer(max.iter), gamma, XG$m, as.integer(dfmax), as.integer(gmax), as.integer(user.lambda))
     b <- rbind(mean(y), matrix(fit[[1]], nrow=p))
-    iter <- fit[[2]]
-    df <- fit[[3]] + 1 # Intercept
-    loss <- fit[[4]]
+    loss <- fit[[2]]
+    Eta <- matrix(fit[[3]], nrow=n) + mean(y)
+    df <- fit[[4]] + 1 # Intercept
+    iter <- fit[[5]]
   } else {
     if (bilevel) fit <- .Call("lcdfit_glm", XG$X, yy, family, penalty, K1, K0, lambda, alpha, eps, 0, gamma, tau, as.integer(max.iter), XG$m, as.integer(dfmax), as.integer(gmax), as.integer(warn), as.integer(user.lambda))
     else fit <- .Call("gdfit_glm", XG$X, yy, family, penalty, K1, K0, lambda, alpha, eps, as.integer(max.iter), gamma, XG$m, as.integer(dfmax), as.integer(gmax), as.integer(warn), as.integer(user.lambda))
     b <- rbind(fit[[1]], matrix(fit[[2]], nrow=p))
-    iter <- fit[[3]]
-    df <- fit[[4]]
-    loss <- fit[[5]]
+    loss <- fit[[3]]
+    Eta <- matrix(fit[[4]], nrow=n)
+    df <- fit[[5]]
+    iter <- fit[[6]]
   }
 
   # Eliminate saturated lambda values, if any
@@ -74,6 +90,7 @@ grpreg <- function(X, y, group=1:ncol(X), penalty=c("grLasso", "grMCP", "grSCAD"
   lambda <- lambda[ind]
   df <- df[ind]
   loss <- loss[ind]
+  Eta <- Eta[,ind]
   if (iter[1] == max.iter) stop("Algorithm failed to converge for any values of lambda.  This indicates a combination of (a) an ill-conditioned feature matrix X and (b) insufficient penalization.  You must fix one or the other for your model to be identifiable.", call.=FALSE)
   if (warn & any(iter==max.iter)) warning("Algorithm failed to converge for all values of lambda", call.=FALSE)
 
@@ -93,6 +110,7 @@ grpreg <- function(X, y, group=1:ncol(X), penalty=c("grLasso", "grMCP", "grSCAD"
   } else {
     dimnames(beta) <- list(varnames, round(lambda, digits=4))
   }
+  colnames(Eta) <- round(lambda, digits=4)
 
   val <- structure(list(beta = beta,
                         family = family,
@@ -100,17 +118,27 @@ grpreg <- function(X, y, group=1:ncol(X), penalty=c("grLasso", "grMCP", "grSCAD"
                         lambda = lambda,
                         alpha = alpha,
                         loss = loss,
+                        linear.predictors = Eta,
                         n = n,
                         penalty = penalty,
                         df = df,
                         iter = iter,
                         group.multiplier = XG$m),
                    class = "grpreg")
-  if (returnX) {
-    val$XG = XG
-    val$y = yy
-  } else if (family=="poisson") {
-    val$y <- y
+  if (family == 'gaussian') {
+    val$y <- yy + attr(yy, 'mean')
+  } else {
+    val$y <- yy
+  }
+  if (returnX) val$XG <- XG
+  if (expanded) {
+    val$meta <- list(knots = knots,
+                     boundary = boundary,
+                     degree = degree,
+                     originalx = originalx,
+                     type = type,
+                     X = X)
+    attr(val, "class") <- c("grpreg", "expanded")
   }
   val
 }