+++ inst/tinytest/basic-functionality.R
@@ -9,22 +9,22 @@ nlam=100
 par(mfcol=c(3,2))
 gel <- coef(fit <- grpreg(X, y, group, penalty="gel", nlambda=nlam, lambda.min=0))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(gel, reg)
+expect_equivalent(gel, reg, tolerance=1e-7)
 cMCP <- coef(fit <- grpreg(X, y, group, penalty="cMCP", nlambda=nlam, lambda.min=0))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(cMCP, reg)
+expect_equivalent(cMCP, reg, tolerance=1e-7)
 bridge <- coef(fit <- gBridge(X, y, group, nlambda=nlam, lambda.min=0))[,1]
 plot(fit, main=fit$penalty)
-expect_equivalent(bridge, reg)
+expect_equivalent(bridge, reg, tolerance=1e-7)
 grLasso <- coef(fit <- grpreg(X, y, group, penalty="grLasso", nlambda=nlam, lambda.min=0))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(grLasso, reg)
+expect_equivalent(grLasso, reg, tolerance=1e-7)
 grMCP <- coef(fit <- grpreg(X, y, group, penalty="grMCP", nlambda=nlam, lambda.min=0))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(grMCP, reg)
+expect_equivalent(grMCP, reg, tolerance=1e-7)
 grSCAD <- coef(fit <- grpreg(X, y, group, penalty="grSCAD", nlambda=nlam, lambda.min=0))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(grSCAD, reg)
+expect_equivalent(grSCAD, reg, tolerance=1e-7)
 
 # grpreg reproduces linear regression
 n <- 50
@@ -38,23 +38,23 @@ nlam=100
 par(mfcol=c(3,2))
 gel <- coef(fit <- grpreg(X, y, group, penalty="gel", nlambda=nlam, lambda.min=0, eps=1e-10))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(gel, reg)
+expect_equivalent(gel, reg, tolerance=1e-7)
 cMCP <- coef(fit <- grpreg(X, y, group, penalty="cMCP", lambda.min=0, eps=1e-10))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(cMCP, reg)
+expect_equivalent(cMCP, reg, tolerance=1e-7)
 bridge <- coef(fit <- gBridge(X, y, group, lambda.min=0, eps=1e-10))[,1]
 plot(fit, main=fit$penalty)
-expect_equivalent(bridge, reg)
+expect_equivalent(bridge, reg, tolerance=1e-7)
 grLasso <- coef(fit <- grpreg(X, y, group, penalty="grLasso", lambda.min=0, eps=1e-10))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(grLasso, reg)
+expect_equivalent(grLasso, reg, tolerance=1e-7)
 grMCP <- coef(fit <- grpreg(X, y, group, penalty="grMCP", lambda.min=0, eps=1e-10))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(grMCP, reg)
+expect_equivalent(grMCP, reg, tolerance=1e-7)
 grSCAD <- coef(fit <- grpreg(X, y, group, penalty="grSCAD", lambda.min=0, eps=1e-10))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(grSCAD, reg)
-expect_equivalent(predict(fit, X)[,100], predict(fit.mle))
+expect_equivalent(grSCAD, reg, tolerance=1e-7)
+expect_equivalent(predict(fit, X)[,100], predict(fit.mle), tolerance=1e-7)
 
 # grpreg reproduces simple logistic regression
 n <- 20
@@ -67,25 +67,25 @@ nlam <- 100
 par(mfcol=c(3,2))
 gel <- coef(fit <- grpreg(X, y, group, penalty="gel", nlambda=nlam, lambda.min=0, family="binomial", eps=1e-10))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(gel, reg)
+expect_equivalent(gel, reg, tolerance=1e-7)
 cMCP <- coef(fit <- grpreg(X, y, group, penalty="cMCP", nlambda=nlam, lambda.min=0, family="binomial", gamma=9, eps=1e-10))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(cMCP, reg)
+expect_equivalent(cMCP, reg, tolerance=1e-7)
 bridge <- coef(fit <- gBridge(X, y, group, lambda.min=0, nlambda=nlam, family="binomial", eps=1e-10))[,1]
 plot(fit, main=fit$penalty)
-expect_equivalent(bridge, reg)
+expect_equivalent(bridge, reg, tolerance=1e-7)
 grLasso <- coef(fit <- grpreg(X, y, group, penalty="grLasso", nlambda=nlam, lambda.min=0, family="binomial", eps=1e-10))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(grLasso, reg)
+expect_equivalent(grLasso, reg, tolerance=1e-7)
 grMCP <- coef(fit <- grpreg(X, y, group, penalty="grMCP", nlambda=nlam, lambda.min=0, family="binomial", eps=1e-10))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(grMCP, reg)
+expect_equivalent(grMCP, reg, tolerance=1e-7)
 grSCAD <- coef(fit <- grpreg(X, y, group, penalty="grSCAD", nlambda=nlam, lambda.min=0, family="binomial", eps=1e-10))[,nlam]
 plot(fit, main=fit$penalty)
-expect_equivalent(grSCAD, reg)
+expect_equivalent(grSCAD, reg, tolerance=1e-7)
 
 # grpreg() reproduces logistic regression
-n <- 50
+n <- 100
 group <- rep(0:3,1:4)
 p <- length(group)
 X <- matrix(rnorm(n*p),ncol=p)
@@ -125,21 +125,21 @@ reg <- coef(fit.mle)
 par(mfcol=c(3,2))
 gel <- coef(fit <- grpreg(X, y, group, penalty="gel", family="poisson", eps=1e-10, lambda.min=0))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(gel, reg)
+expect_equivalent(gel, reg, tolerance=1e-7)
 cMCP <- coef(fit <- grpreg(X, y, group, penalty="cMCP", family="poisson", eps=1e-10, lambda.min=0))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(cMCP, reg)
+expect_equivalent(cMCP, reg, tolerance=1e-7)
 bridge <- coef(fit <- gBridge(X, y, group, family="poisson", eps=1e-10, lambda.min=0))[,1]
 plot(fit, main=fit$penalty)
-expect_equivalent(bridge, reg)
+expect_equivalent(bridge, reg, tolerance=1e-7)
 grLasso <- coef(fit <- grpreg(X, y, group, penalty="grLasso", family="poisson", eps=1e-10, lambda.min=0))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(grLasso, reg)
+expect_equivalent(grLasso, reg, tolerance=1e-7)
 grMCP <- coef(fit <- grpreg(X, y, group, penalty="grMCP", family="poisson", gamma=2, eps=1e-10, lambda.min=0))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(grMCP, reg)
+expect_equivalent(grMCP, reg, tolerance=1e-7)
 grSCAD <- coef(fit <- grpreg(X, y, group, penalty="grSCAD", family="poisson", gamma=2.1, eps=1e-10, lambda.min=0))[,100]
 plot(fit, main=fit$penalty)
-expect_equivalent(grSCAD, reg)
-expect_equivalent(predict(fit, X)[,100], predict(fit.mle))
-expect_equivalent(predict(fit, X, type="response")[,100], predict(fit.mle, type="response"))
+expect_equivalent(grSCAD, reg, tolerance=1e-7)
+expect_equivalent(predict(fit, X)[,100], predict(fit.mle), tolerance=1e-7)
+expect_equivalent(predict(fit, X, type="response")[,100], predict(fit.mle, type="response"), tolerance=1e-7)